fix(wasmcloud): instantiate using the proper linker
