+++ crates/wasmcloud/tests/integration_http_blobby.rs
@@ -72,6 +72,7 @@ async fn test_blobby_integration() -> Result<()> {
                     memory_limit_mb: 256,
                     cpu_limit: 1,
                     config: HashMap::new(),
+                    environment: HashMap::new(),
                     volume_mounts: vec![],
                     allowed_hosts: vec![],
                 },
@@ -234,6 +235,7 @@ async fn test_blobby_error_handling() -> Result<()> {
                     memory_limit_mb: 128,
                     cpu_limit: 1,
                     config: HashMap::new(),
+                    environment: HashMap::new(),
                     volume_mounts: vec![],
                     allowed_hosts: vec![],
                 },
