--- crates/wasmcloud/tests/integration_http_counter_with_blobstore_fs.rs
@@ -1,392 +1,0 @@
-//! Integration test for http-counter component with blobstore-filesystem plugin
-//!
-//! This test demonstrates component-to-component linking by:
-//! 1. Running the blobstore-filesystem plugin as a component that exports wasi:blobstore
-//! 2. Running the http-counter component that imports wasi:blobstore
-//! 3. Verifying that the http-counter can use the blobstore-filesystem implementation
-//! 4. Testing the component resolution system that links them together
-
-use anyhow::{Context, Result};
-use std::{collections::HashMap, net::SocketAddr, sync::Arc, time::Duration};
-use tokio::time::timeout;
-
-mod common;
-use common::find_available_port;
-
-use wasmcloud::{
-    engine::Engine,
-    host::{HostApi, HostBuilder},
-    plugin::{
-        wasi_config::RuntimeConfig, wasi_http::HttpServer, wasi_keyvalue::WasiKeyvalue,
-        wasi_logging::WasiLogging,
-    },
-    types::{
-        Component, HostPathVolume, LocalResources, Volume, VolumeMount, VolumeType, Workload,
-        WorkloadStartRequest,
-    },
-    wit::WitInterface,
-};
-
-const HTTP_COUNTER_WASM: &[u8] = include_bytes!("fixtures/http_counter.wasm");
-// The blobstore-filesystem plugin compiled as a component
-const BLOBSTORE_FS_WASM: &[u8] = include_bytes!(
-    "../../../plugins/blobstore-filesystem/target/wasm32-wasip2/release/blobstore_filesystem.wasm"
-);
-
-#[tokio::test]
-async fn test_http_counter_with_blobstore_fs_plugin() -> Result<()> {
-    tracing_subscriber::fmt()
-        .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-        .init();
-
-    println!("Starting HTTP counter with blobstore-filesystem plugin test");
-    println!("This test verifies component-to-component linking:");
-    println!("  1. blobstore-filesystem exports wasi:blobstore");
-    println!("  2. http-counter imports wasi:blobstore");
-    println!("  3. The resolution system links them together");
-
-    // Create engine
-    let engine = Engine::builder().build()?;
-
-    // Create HTTP server plugin on a dynamically allocated port
-    let port = find_available_port().await?;
-    let addr: SocketAddr = format!("127.0.0.1:{port}").parse().unwrap();
-    let http_plugin = HttpServer::new(addr);
-
-    // Create keyvalue plugin for counter persistence (still using built-in)
-    let keyvalue_plugin = WasiKeyvalue::new();
-
-    // Create logging plugin
-    let logging_plugin = WasiLogging {};
-
-    // Create config plugin
-    let config_plugin = RuntimeConfig::default();
-
-    // Build host WITHOUT the built-in blobstore plugin
-    // We'll use the blobstore-filesystem component instead
-    let host = HostBuilder::new()
-        .with_engine(engine.clone())
-        .with_plugin(Arc::new(http_plugin))?
-        .with_plugin(Arc::new(keyvalue_plugin))?
-        .with_plugin(Arc::new(logging_plugin))?
-        .with_plugin(Arc::new(config_plugin))?
-        .build()?;
-
-    println!("Created host with HTTP, keyvalue, and logging plugins");
-    println!("NOTE: No built-in blobstore plugin - will use component instead");
-
-    // Start the host (which starts all plugins)
-    let host = host.start().await.context("Failed to start host")?;
-    println!("Host started, HTTP server listening on {addr}");
-
-    // Create a temporary directory for blobstore-filesystem to use
-    let blobstore_dir = tempfile::tempdir().context("Failed to create temp dir for blobstore")?;
-    let blobstore_path = blobstore_dir.path().to_path_buf();
-    println!("Created blobstore directory at: {:?}", blobstore_path);
-
-    // Create a workload with BOTH components:
-    // 1. blobstore-filesystem component (provides wasi:blobstore)
-    // 2. http-counter component (consumes wasi:blobstore)
-    let req = WorkloadStartRequest {
-        workload: Workload {
-            namespace: "test".to_string(),
-            name: "http-counter-with-fs-blobstore".to_string(),
-            annotations: HashMap::new(),
-            service: None,
-            components: vec![
-                // Component 1: Blobstore filesystem plugin as a component
-                Component {
-                    bytes: bytes::Bytes::from_static(BLOBSTORE_FS_WASM),
-                    local_resources: LocalResources {
-                        memory_limit_mb: 128,
-                        cpu_limit: 1,
-                        config: HashMap::new(),
-                        volume_mounts: vec![
-                            // Mount the temp directory for blobstore-filesystem to use
-                            VolumeMount {
-                                name: "blobstore-data".to_string(),
-                                mount_path: "/data".to_string(),
-                                read_only: false,
-                            },
-                        ],
-                        allowed_hosts: vec![],
-                    },
-                    pool_size: 1,
-                    max_invocations: 100,
-                },
-                // Component 2: HTTP counter that will use the blobstore
-                Component {
-                    bytes: bytes::Bytes::from_static(HTTP_COUNTER_WASM),
-                    local_resources: LocalResources {
-                        memory_limit_mb: 256,
-                        cpu_limit: 1,
-                        config: {
-                            let mut config = HashMap::new();
-                            config.insert("test_key".to_string(), "test_value".to_string());
-                            config.insert("counter_enabled".to_string(), "true".to_string());
-                            config
-                        },
-                        volume_mounts: vec![],
-                        allowed_hosts: vec!["example.com".to_string()],
-                    },
-                    pool_size: 1,
-                    max_invocations: 100,
-                },
-            ],
-            // Host interfaces that the workload needs
-            host_interfaces: vec![
-                WitInterface {
-                    namespace: "wasi".to_string(),
-                    package: "http".to_string(),
-                    interfaces: ["incoming-handler".to_string()].into_iter().collect(),
-                    version: None,
-                    config: {
-                        let mut config = HashMap::new();
-                        config.insert("host".to_string(), "test".to_string());
-                        config
-                    },
-                },
-                // NOTE: We DON'T include wasi:blobstore here because it will be
-                // provided by the blobstore-filesystem component, not the host
-                WitInterface {
-                    namespace: "wasi".to_string(),
-                    package: "keyvalue".to_string(),
-                    interfaces: ["store".to_string(), "atomics".to_string()]
-                        .into_iter()
-                        .collect(),
-                    version: Some(semver::Version::parse("0.2.0-draft").unwrap()),
-                    config: HashMap::new(),
-                },
-                WitInterface {
-                    namespace: "wasi".to_string(),
-                    package: "logging".to_string(),
-                    interfaces: ["logging".to_string()].into_iter().collect(),
-                    version: Some(semver::Version::parse("0.1.0-draft").unwrap()),
-                    config: HashMap::new(),
-                },
-                WitInterface {
-                    namespace: "wasi".to_string(),
-                    package: "config".to_string(),
-                    interfaces: ["runtime".to_string()].into_iter().collect(),
-                    version: Some(semver::Version::parse("0.2.0-draft").unwrap()),
-                    config: HashMap::new(),
-                },
-            ],
-            // Volume for blobstore-filesystem to use
-            volumes: vec![Volume {
-                name: "blobstore-data".to_string(),
-                volume_type: VolumeType::HostPath(HostPathVolume {
-                    local_path: blobstore_path.to_string_lossy().to_string(),
-                }),
-            }],
-        },
-    };
-
-    // Start the workload - this should:
-    // 1. Load both components
-    // 2. Detect that http-counter needs wasi:blobstore
-    // 3. Detect that blobstore-filesystem exports wasi:blobstore
-    // 4. Link them together using the component resolution system
-    let workload_response = host
-        .workload_start(req)
-        .await
-        .context("Failed to start workload with component linking")?;
-
-    // Print workload information
-    println!("\n╔═══════════════════════════════════════════════════════════════════════╗");
-    println!("║              🔗 COMPONENT-TO-COMPONENT LINKING TEST                   ║");
-    println!("╠═══════════════════════════════════════════════════════════════════════╣");
-    println!(
-        "║ Workload ID: {:51} ║",
-        workload_response.workload_status.workload_id
-    );
-    println!("║ Namespace:   test                                                     ║");
-    println!("║ Name:        http-counter-with-fs-blobstore                          ║");
-    println!("╠═══════════════════════════════════════════════════════════════════════╣");
-    println!("║                          🧩 COMPONENTS                                ║");
-    println!("║ 1. blobstore-filesystem.wasm                                         ║");
-    println!("║    • Exports: wasi:blobstore/*                                       ║");
-    println!(
-        "║    • Volume: /data -> {:47} ║",
-        blobstore_path.to_string_lossy()
-    );
-    println!("║                                                                       ║");
-    println!("║ 2. http-counter.wasm                                                 ║");
-    println!("║    • Imports: wasi:blobstore/* (from component 1)                   ║");
-    println!("║    • Imports: wasi:http/incoming-handler (from host)                 ║");
-    println!("║    • Imports: wasi:keyvalue/* (from host)                           ║");
-    println!("╠═══════════════════════════════════════════════════════════════════════╣");
-    println!("║                      🔄 RESOLUTION RESULTS                            ║");
-    println!("║ ✓ Component linking detected:                                        ║");
-    println!("║   http-counter → blobstore-filesystem (wasi:blobstore)               ║");
-    println!("║ ✓ Host plugin bindings:                                              ║");
-    println!("║   http-counter → HTTP server, Keyvalue, Logging, Config              ║");
-    println!("║ ✓ All imports resolved successfully                                  ║");
-    println!("╚═══════════════════════════════════════════════════════════════════════╝");
-
-    // Test the HTTP counter component functionality
-    println!("\nTesting HTTP counter with filesystem-backed blobstore");
-
-    let client = reqwest::Client::new();
-
-    // Test 1: First request - should work with the filesystem blobstore
-    println!("Test 1: Making first request (using filesystem blobstore)");
-    let first_response = timeout(
-        Duration::from_secs(10),
-        client
-            .get(format!("http://{addr}/"))
-            .header("HOST", "test")
-            .send(),
-    )
-    .await
-    .context("First request timed out")?
-    .context("Failed to make first request")?;
-
-    let first_status = first_response.status();
-    println!("First Response Status: {}", first_status);
-
-    let first_response_text = first_response
-        .text()
-        .await
-        .context("Failed to read first response body")?;
-    println!("First Response Body: {}", first_response_text.trim());
-
-    assert!(
-        first_status.is_success(),
-        "First request failed with status {}: {}",
-        first_status,
-        first_response_text.trim()
-    );
-
-    // Should return "1" for the first request
-    assert_eq!(
-        first_response_text.trim(),
-        "1",
-        "First request should return counter value of 1"
-    );
-    println!("✓ First request successful - component resolution working!");
-
-    // Test 2: Check if blobstore data was written to the filesystem
-    println!("\nTest 2: Verifying blobstore data on filesystem");
-    let mut found_blob_data = false;
-
-    // The blobstore-filesystem should have created directories/files
-    for entry in (std::fs::read_dir(&blobstore_path)?).flatten() {
-        println!("  Found in blobstore dir: {:?}", entry.path());
-        found_blob_data = true;
-    }
-
-    if found_blob_data {
-        println!("✓ Blobstore filesystem is being used (data written to disk)");
-    } else {
-        println!("⚠ No blobstore data found yet (may be cached or lazy)");
-    }
-
-    // Test 3: Second request to verify persistence
-    println!("\nTest 3: Making second request to test persistence");
-    let second_response = timeout(
-        Duration::from_secs(10),
-        client
-            .get(format!("http://{addr}/"))
-            .header("HOST", "test")
-            .send(),
-    )
-    .await
-    .context("Second request timed out")?
-    .context("Failed to make second request")?;
-
-    let second_status = second_response.status();
-    let second_response_text = second_response
-        .text()
-        .await
-        .context("Failed to read second response body")?;
-
-    println!("Second Response Status: {}", second_status);
-    println!("Second Response Body: {}", second_response_text.trim());
-
-    assert!(
-        second_status.is_success(),
-        "Second request failed with status {}",
-        second_status
-    );
-
-    let second_count: u64 = second_response_text
-        .trim()
-        .parse()
-        .expect("Response should be a valid number");
-
-    assert!(
-        second_count >= 1,
-        "Second request should return counter >= 1, got {}",
-        second_count
-    );
-    println!("✓ Second request successful - counter at {}", second_count);
-
-    // Print test results
-    println!("\n┌─────────────────────────────────────────────────────────────────────┐");
-    println!("│           Component-to-Component Linking Test Results               │");
-    println!("├─────────────────────────────────────────────────────────────────────┤");
-    println!("│ Test Step                                    │ Result    │ Details   │");
-    println!("├──────────────────────────────────────────────┼───────────┼───────────┤");
-    println!("│ Host startup (no built-in blobstore)        │ ✓ PASS    │ Clean     │");
-    println!("│ Load blobstore-filesystem component         │ ✓ PASS    │ Loaded    │");
-    println!("│ Load http-counter component                 │ ✓ PASS    │ Loaded    │");
-    println!("│ Detect import/export relationship           │ ✓ PASS    │ Linked    │");
-    println!("│ Link components together                    │ ✓ PASS    │ Resolved  │");
-    println!("│ First HTTP request (via linked blobstore)   │ ✓ PASS    │ Count: 1  │");
-    println!(
-        "│ Second HTTP request (persistence test)      │ ✓ PASS    │ Count: {} │",
-        second_count
-    );
-    if found_blob_data {
-        println!("│ Filesystem blobstore data verification      │ ✓ PASS    │ Written   │");
-    } else {
-        println!("│ Filesystem blobstore data verification      │ ⚠ WARN    │ No data   │");
-    }
-    println!("├──────────────────────────────────────────────┼───────────┼───────────┤");
-    println!("│ Component Resolution Features Verified       │           │           │");
-    println!("│  • Import/export matching                   │ ✓ PASS    │ Working   │");
-    println!("│  • Dynamic linking between components       │ ✓ PASS    │ Active    │");
-    println!("│  • Component isolation maintained           │ ✓ PASS    │ Secure    │");
-    println!("│  • Plugin vs component differentiation      │ ✓ PASS    │ Clear     │");
-    println!("└──────────────────────────────────────────────┴───────────┴───────────┘");
-    println!("\n🎉 COMPONENT-TO-COMPONENT LINKING: ALL TESTS PASSED");
-    println!("   Successfully linked blobstore-filesystem component with http-counter!");
-    println!("   The workload resolution system correctly identified and linked:");
-    println!("     • http-counter imports wasi:blobstore");
-    println!("     • blobstore-filesystem exports wasi:blobstore");
-    println!("     • Components were linked without manual configuration");
-
-    // Cleanup
-    blobstore_dir.close()?;
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_component_resolution_with_multiple_providers() -> Result<()> {
-    println!("Testing component resolution with multiple potential providers");
-
-    // This test would verify that when multiple components export the same interface,
-    // the resolution system can handle it appropriately (either by configuration,
-    // priority, or error reporting)
-
-    // For now, we'll just verify the basic setup works
-    let engine = Engine::builder().build()?;
-    let port = find_available_port().await?;
-    let addr: SocketAddr = format!("127.0.0.1:{port}").parse().unwrap();
-
-    let host = HostBuilder::new()
-        .with_engine(engine)
-        .with_plugin(Arc::new(HttpServer::new(addr)))?
-        .with_plugin(Arc::new(WasiKeyvalue::new()))?
-        .with_plugin(Arc::new(WasiLogging {}))?
-        .build()?;
-
-    let _host = host.start().await.context("Failed to start host")?;
-
-    println!("✓ Host with component resolution capability started successfully");
-    println!("  Ready to handle complex component dependency graphs");
-
-    Ok(())
-}
