+++ crates/wasmcloud/tests/integration_http_keyvalue.rs
@@ -73,6 +73,7 @@ async fn test_http_keyvalue_counter_integration() -> Result<()> {
                     memory_limit_mb: 256,
                     cpu_limit: 1,
                     config: HashMap::new(),
+                    environment: HashMap::new(),
                     volume_mounts: vec![],
                     allowed_hosts: vec![],
                 },
@@ -322,6 +323,7 @@ async fn test_keyvalue_counter_concurrent_access() -> Result<()> {
                     memory_limit_mb: 256,
                     cpu_limit: 1,
                     config: HashMap::new(),
+                    environment: HashMap::new(),
                     volume_mounts: vec![],
                     allowed_hosts: vec![],
                 },
@@ -482,6 +484,7 @@ async fn test_keyvalue_error_handling() -> Result<()> {
                     memory_limit_mb: 128,
                     cpu_limit: 1,
                     config: HashMap::new(),
+                    environment: HashMap::new(),
                     volume_mounts: vec![],
                     allowed_hosts: vec![],
                 },
